bisync: factor resync to separate file
